+++ scripts/templates/template.py.jinja2
@@ -47,6 +47,7 @@ class {{ message.name }}(MessageMarshaler):
 {%- for field in message.fields %}
         self.{{ field.attr_name }}: {{ field.python_type }} = {{ field.name }}
 {%- endfor %}
+        self._marshaler_cache = {}
 
     def calculate_size(self) -> int:
         size = 0
