+++ src/snowflake/telemetry/_internal/opentelemetry/proto/common/v1/common_marshaler.py
@@ -54,12 +54,12 @@ class AnyValue(MessageMarshaler):
         self._array_value: ArrayValue = array_value
         self._kvlist_value: KeyValueList = kvlist_value
         self.bytes_value: bytes = bytes_value
+        self._marshaler_cache = {}
 
     def calculate_size(self) -> int:
         size = 0
         if self.string_value is not None:
             v = self.string_value.encode("utf-8")
-            self._string_value_encoded = v
             size += len(b"\n") + Varint.size_varint_u32(len(v)) + len(v)
         if self.bool_value is not None:
             size += len(b"\x10") + 1
@@ -89,7 +89,7 @@ class AnyValue(MessageMarshaler):
 
     def write_to(self, out: BytesIO) -> None:
         if self.string_value is not None:
-            v = self._string_value_encoded
+            v = self.string_value.encode("utf-8")
             out += b"\n"
             Varint.write_varint_u32(out, len(v))
             out += v
@@ -128,6 +128,7 @@ class ArrayValue(MessageMarshaler):
         values: List[AnyValue] = None,
     ):
         self._values: List[AnyValue] = values
+        self._marshaler_cache = {}
 
     def calculate_size(self) -> int:
         size = 0
@@ -160,6 +161,7 @@ class KeyValueList(MessageMarshaler):
         values: List[KeyValue] = None,
     ):
         self._values: List[KeyValue] = values
+        self._marshaler_cache = {}
 
     def calculate_size(self) -> int:
         size = 0
@@ -196,12 +198,12 @@ class KeyValue(MessageMarshaler):
     ):
         self.key: str = key
         self._value: AnyValue = value
+        self._marshaler_cache = {}
 
     def calculate_size(self) -> int:
         size = 0
         if self.key:
             v = self.key.encode("utf-8")
-            self._key_encoded = v
             size += len(b"\n") + Varint.size_varint_u32(len(v)) + len(v)
         if self._value is not None:
             size += (
@@ -213,7 +215,7 @@ class KeyValue(MessageMarshaler):
 
     def write_to(self, out: BytesIO) -> None:
         if self.key:
-            v = self._key_encoded
+            v = self.key.encode("utf-8")
             out += b"\n"
             Varint.write_varint_u32(out, len(v))
             out += v
@@ -246,16 +248,15 @@ class InstrumentationScope(MessageMarshaler):
         self.version: str = version
         self._attributes: List[KeyValue] = attributes
         self.dropped_attributes_count: int = dropped_attributes_count
+        self._marshaler_cache = {}
 
     def calculate_size(self) -> int:
         size = 0
         if self.name:
             v = self.name.encode("utf-8")
-            self._name_encoded = v
             size += len(b"\n") + Varint.size_varint_u32(len(v)) + len(v)
         if self.version:
             v = self.version.encode("utf-8")
-            self._version_encoded = v
             size += len(b"\x12") + Varint.size_varint_u32(len(v)) + len(v)
         if self._attributes:
             size += sum(
@@ -270,12 +271,12 @@ class InstrumentationScope(MessageMarshaler):
 
     def write_to(self, out: BytesIO) -> None:
         if self.name:
-            v = self._name_encoded
+            v = self.name.encode("utf-8")
             out += b"\n"
             Varint.write_varint_u32(out, len(v))
             out += v
         if self.version:
-            v = self._version_encoded
+            v = self.version.encode("utf-8")
             out += b"\x12"
             Varint.write_varint_u32(out, len(v))
             out += v
