+++ src/snowflake/telemetry/_internal/opentelemetry/proto/resource/v1/resource_marshaler.py
@@ -34,6 +34,7 @@ class Resource(MessageMarshaler):
     ):
         self._attributes: List[KeyValue] = attributes
         self.dropped_attributes_count: int = dropped_attributes_count
+        self._marshaler_cache = {}
 
     def calculate_size(self) -> int:
         size = 0
